feat(security): add get_encryption_key with fail-closed production security
Add AES-256 encryption key retrieval from environment variables with fail-closed
security model. In production mode, missing JARVIS_ENCRYPTION_KEY halts boot to
prevent plaintext memory exposure. Training mode allows plaintext with warning.

Also includes:
- Fix ENV lookup bug in RustIPC.jl load_key_from_environment()
- Add Float32 type annotations for candidate scores in Kernel.jl and run_kernel.jl
- Update Project.toml with MbedTLS dependency for TLS support

+++ adaptive-kernel/kernel/Kernel.jl
@@ -745,7 +745,7 @@ function select_action(kernel::KernelState, capability_candidates::Vector{<:Any}
         get(chosen, "id", "unknown"),
         Float32(get(chosen, "confidence", 0.7)),
         Float32(get(chosen, "cost", 0.1)),
-        best_score,
+        Float32(best_score),
         risk_value,
         "Selected by priority heuristic (score=$(round(best_score, digits=3)))"
     )
